refactor(web-serial-data-access): serial services for rxjs v2.3.1 (#606)
