--- libs/web-serial/data-access/src/lib/serial-command.service.ts
@@ -1,2 +1,0 @@
-export type { CommandExecutionConfig, CommandResult } from './serial-command/serial-command-facade.service';
-export { SerialCommandService } from './serial-command/serial-command-facade.service';
